pyproject.toml
[project]
name = "workshop-agents"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "basedpyright>=1.39.0",
    "deepeval>=3.9.6",
    "fastapi>=0.135.3",
    "google-cloud-aiplatform>=1.147.0",
    "google-genai>=1.71.0",
    "langfuse>=4.0.6",
    "loguru>=0.7.3",
    "opentelemetry-exporter-otlp>=1.40.0",
    "opentelemetry-sdk>=1.40.0",
    "pydantic-ai>=0.8.1",
    "python-dotenv>=1.2.2",
    "ruff>=0.15.10",
    "uvicorn>=0.44.0",
]
